consumption temp

+++ src/main/java/com/jandagort/game/economy/planet/Production.java
@@ -4,7 +4,7 @@ import lombok.Data;
 import org.springframework.context.annotation.Scope;
 
 import javax.persistence.*;
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Scope("prototype")
 @Data
@@ -14,7 +14,7 @@ public class Production {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private long id;
-    private BigInteger food;
-    private BigInteger electricity;
-    private BigInteger residence;
+    private BigDecimal food;
+    private BigDecimal electricity;
+    private BigDecimal residence;
 }
