+++ src/main/java/com/jandagort/register/RegisterConfiguration.java
@@ -8,7 +8,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Component
 public class RegisterConfiguration {
@@ -31,9 +31,9 @@ public class RegisterConfiguration {
     @Scope("prototype")
     public Supply supply() {
         Supply supply = new Supply();
-        supply.setFood(new BigInteger("0"));
-        supply.setElectricity(new BigInteger("0"));
-        supply.setPopulation(new BigInteger("0"));
+        supply.setFood(new BigDecimal("0"));
+        supply.setElectricity(new BigDecimal("0"));
+        supply.setPopulation(new BigDecimal("0"));
 
         return supply;
     }
@@ -43,9 +43,9 @@ public class RegisterConfiguration {
     @Scope("prototype")
     public Consumption consumption() {
         Consumption consumption = new Consumption();
-        consumption.setFood(new BigInteger("0"));
-        consumption.setElectricity(new BigInteger("0"));
-        consumption.setPopulation(new BigInteger("0"));
+        consumption.setFood(new BigDecimal("0"));
+        consumption.setElectricity(new BigDecimal("0"));
+        consumption.setPopulation(new BigDecimal("0"));
 
         return consumption;
     }
@@ -55,9 +55,9 @@ public class RegisterConfiguration {
     @Scope("prototype")
     public Production resources() {
         Production production = new Production();
-        production.setFood(new BigInteger("0"));
-        production.setElectricity(new BigInteger("0"));
-        production.setResidence(new BigInteger("0"));
+        production.setFood(new BigDecimal("0"));
+        production.setElectricity(new BigDecimal("0"));
+        production.setResidence(new BigDecimal("0"));
 
         return production;
     }
